requirements.txt
pyspark
grpcio-status
zstandard
yfinance
streamlit
cassandra-driver
prophet